BUG FIX: there was a bug when converting from long to int.

+++ trunk/comaWebServer/src/main/java/bioinfo/comaWebServer/util/JobGridDataSource.java
@@ -39,7 +39,7 @@ public class JobGridDataSource implements GridDataSource
 		
 		transaction.commit();
 
-		return (Integer)list.get(0);
+		return ((Long)list.get(0)).intValue();
 	}
 
 	@Override
